fix：模块化PFC
